% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afhouse}
\alias{afhouse}
\title{affordable rental housing projects}
\format{
dataframe with 428 observations of 14 variables
\describe{
  \item{}{}
}
}
\source{
\url{https://data.cityofchicago.org/Community-Economic-Development/Affordable-Rental-Housing-Developments/s6ha-ppgi/data}
}
\usage{
afhouse
}
\description{
affordable rental housing projects
}
\details{
csv file with coordinate informations for each project
}
\keyword{datasets}
