% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca_boundaries}
\alias{ca_boundaries}
\title{sf object of Chicago community areas}
\format{An sf object with 77 rows and 6 variables:
\describe{
  \item{area_num_1}{id number for CA}
  \item{commnity}{historic name of community}
  \item{perimeter}{}
  \item{shape_area}{geometry}
  \item{geometry}{sfc polygon,  crs = 4362}
}}
\source{
\url{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Community-Areas-current-/cauq-8yn6}
}
\usage{
ca_boundaries
}
\description{
sf object of Chicago community areas
}
\details{
not much to say here, just polygons with neighborhood names and other
identifying information. Current since 2013.
}
\keyword{datasets}
