% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_tracts_2010}
\alias{census_tracts_2010}
\title{boundary files of 2010 census tracrs in the city of Chicago}
\format{
sf object (st_polygons) with 798 observations of 28 variables
\describe{
  \item{}{}
}
}
\source{
\url{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Census-Tracts-2010/5jrd-6zik}
}
\usage{
census_tracts_2010
}
\description{
boundary files of 2010 census tracrs in the city of Chicago
}
\details{
census tract boundaries
}
\keyword{datasets}
