% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chicago_svi}
\alias{chicago_svi}
\title{Cleaned sf SVI data on tract level}
\format{An sf object with 801 rows and 14 variables:
\describe{
  \item{tractce10}{tract ID in numeric format (had to do some cleaning that made it no longer a 10-13 digit pin}
  \item{commarea}{numeric id of community area tract is in}
  \item{namelsad10}{?}
  \item{E_HU}{housing units estimate}
  \item{HH}{estimate of #of households}
  \item{E_POV}{number persons below poverty estimate}
  \item{E_UNEMP}{number civilians 16+ that are unemployed}
  \item{E_PCI}{per capita income estimate}
  \item{M_PCI}{margin of error on per capita income}
  \item{E_NOHSDIP}{persons 25+ that have no HS diploma}
  \item{MINORITY}{all persons except white, non hispanic, estimate}
  \item{E_LIMENG}{persons 5+ who speak Enligsh "less than well"}
  \item{R_PL_THEME1}{percentile ranking for socioeconomic theme summary - higher value is higher vulnerability}
  \item{geometry}{sfc polygon, crs = 4326}
}}
\source{
\url{https://svi.cdc.gov/data-and-tools-download.html}
}
\usage{
chicago_svi
}
\description{
Cleaned sf SVI data on tract level
}
\details{
A sf dataset containing selected variables from the CDCs Social Vulnerability Index.
Data is from the American Community Survey averaged between 2012-2016 and, accordingly, pure #s in variables
are estimated from this. I cleaned this down to a collection of variables that I thought
provided a good summary, but the original data is in data-raw with basic cleaning methods included.

the documentation for this dataset is extensive:
https://svi.cdc.gov/Documents/Data/2016_SVI_Data/SVI2016Documentation.pdf
}
\keyword{datasets}
