% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cps_tiers}
\alias{cps_tiers}
\title{sf object of Chicago Public School Tiers}
\format{
798 observations of 28 variables
\describe{
  \item{tractce10}{}
  \item{}{}
}
}
\source{
\url{}
}
\usage{
cps_tiers
}
\description{
sf object of Chicago Public School Tiers
}
\details{
A sf object with information regarding the 798 public schools in Chicago. Information mainly concerns schools' tiers from 2011 to 2018.
}
\keyword{datasets}
