% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpsPR2018_sf}
\alias{cpsPR2018_sf}
\title{2018-2019 CPS Progress report sf object}
\format{An sf object with 660 rows and 55 variables:
\describe{
  \item{School_ID}{}
  \item{Short_Name}{}
  \item{Long_Name}{}
  ...
  \item{geometry}{sfc points, crs = 4362}
}}
\source{
\url{https://data.cityofchicago.org/Education/Chicago-Public-Schools-School-Progress-Reports-SY1/dw27-rash}
}
\usage{
cpsPR2018_sf
}
\description{
A sf dataset containing the a selection of variables from CPS annual progress report.
}
\details{
Schools have been geolocated, with data (mostly) cleaned. This was a challenging dataset because of CPS
extremely sparse documentation
}
\keyword{datasets}
