% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_sf.R
\name{csv_to_sf}
\alias{csv_to_sf}
\title{converts a csv file with geometry to sf object}
\usage{
csv_to_sf(name, dataset, long_name, lat_name, espg)
}
\arguments{
\item{name}{you want for the sf object (string).}

\item{dataset}{path for dataset being transformed (string).}

\item{long_name}{name of the longitude column from your file (string).}

\item{lat_name}{name of latitude column from your file (string).}

\item{espg}{espg of your dataset (integer).}
}
\value{
sf object
}
\description{
converts a csv file with geometry to sf object
}
\examples{
csv_to_sf("affordable_housing",
"~/Documents/code/chicagodatapackage/chicagodatapackage/data-raw/AffordableHousingDevelopments.csv",
"Longitude", "Latitude", 4326)

}
