% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cta_ridership}
\alias{cta_ridership}
\title{cta ridership information}
\format{
dataframe with 30468 observations of 7 variables
\describe{
  \item{}{}
}
}
\source{
\url{https://data.cityofchicago.org/Transportation/CTA-Ridership-Bus-Routes-Monthly-Day-Type-Averages/bynn-gwxy}
}
\usage{
cta_ridership
}
\description{
cta ridership information
}
\details{
Ridership information for all L train stations with monthly and weekday/weekend information.
}
\keyword{datasets}
