% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cta_stations}
\alias{cta_stations}
\title{L train stations}
\format{

}
\source{
\url{https://data.cityofchicago.org/Transportation/CTA-L-Rail-Stations-Shapefile/vmyy-m9qj}
}
\usage{
cta_stations
}
\description{
L train stations
}
\details{
sf object with CTA L train station locations
}
\keyword{datasets}
