% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{divvy}
\alias{divvy}
\title{sf object of divvy bike stations in Chicago}
\format{An sf object with 608 rows and 13 variables:
\describe{
  \item{ID}{station ID}
  \item{Station.Name}{}
  \item{Address}{}
  \item{Total.Docks}{}
  \item{Docks.in.Service}{}
  \item{Status}{}
  \item{Location}{coordinates of station in (long, lat) format}
  \item{Historical.Wards.2003.2015}{}
  \item{Zip.Codes}{}
  \item{Community.Areas}{}
  \item{Census.Tracts}{}
  \item{Wards}{}
  \item{geometry}{sfc points,  crs = 4362}
}}
\source{
\url{https://data.cityofchicago.org/Transportation/Divvy-Bicycle-Stations/bbyy-e7gq}
}
\usage{
divvy
}
\description{
sf object of divvy bike stations in Chicago
}
\details{
A sf dataset containing information on each divvy bike station in Chicago metro area.
Updated Daily, most recent downloaded by me on May 13, it has daily information on each station
— could be fun/interesting to play with sequencing.
}
\keyword{datasets}
