% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{police_beats2019}
\alias{police_beats2019}
\title{sf object of current CPD beats}
\format{An sf object with 277 rows and 5 variables:
\describe{
  \item{beat}{beat ID #}
  \item{beat_num}{}
  \item{district}{police district}
  \item{sector}{}
  \item{geometry}{sfc polygon,  crs = 4362}
}}
\source{
\url{https://data.cityofchicago.org/Public-Safety/Boundaries-Police-Beats-current-/aerh-rz74}
}
\usage{
police_beats2019
}
\description{
sf object of current CPD beats
}
\details{
A sf dataset containing polygons and basic administrative info on police beats in Chicago.
}
\keyword{datasets}
