% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tif_sf}
\alias{tif_sf}
\title{sf object of Tax Increment Financing (TIF) Projects (points)}
\format{
sf object with 502 observations of 19 variables
\describe{
  \item{}{}
}
}
\source{
\url{https://data.cityofchicago.org/Community-Economic-Development/Tax-Increment-Financing-TIF-Projects-Map/v3a3-hhqn}
}
\usage{
tif_sf
}
\description{
sf object of Tax Increment Financing (TIF) Projects (points)
}
\keyword{datasets}
