% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wards2019}
\alias{wards2019}
\title{sf object of the current Chicago wards}
\format{An sf object with 50 rows and 4 variables:
\describe{
  \item{shape_area}{area of ward in ? measurement}
  \item{ward}{id of ward}
  \item{geometry}{sfc multipolygon, crs = 4362}
}}
\source{
\url{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Wards-2015-/sp34-6z76}
}
\usage{
wards2019
}
\description{
sf object of the current Chicago wards
}
\details{
an sf dataset containing the geometry of wards in Chicago since 2015
(the last time a new city council was sworn in). Includes area, ID, and geometry.
}
\keyword{datasets}
